feat: CreateHomeworkModal 폴더 이동 및 내부 로직 수정 #285
